+++ notes.txt
@@ -114,4 +114,4 @@ add draft optimization
 3199
 
 invariant
-3206
+3094
