+++ src/internal.ts
@@ -1,10 +1,10 @@
-export * from "./env"
-export * from "./errors"
-export * from "./types-external"
-export * from "./types-internal"
-export * from "./common"
-export * from "./plugins"
-export * from "./scope"
-export * from "./finalize"
-export * from "./proxy"
-export * from "./immerClass"
+export * from "./crap/env"
+export * from "./crap/errors"
+export * from "./types/types-external"
+export * from "./types/types-internal"
+export * from "./crap/common"
+export * from "./crap/plugins"
+export * from "./core/scope"
+export * from "./core/finalize"
+export * from "./core/proxy"
+export * from "./core/immerClass"
