+++ tsconfig.json
@@ -15,5 +15,5 @@
     "module": "esnext",
     "allowJs": true
   },
-  "files": ["./src/immer.ts", "./src/plugins/es5.ts", "./src/plugins/mapset.ts", "./src/plugins/patches.ts", "./src/globals.d.ts"]
+  "files": ["./src/immer.ts", "./src/plugins/es5.ts", "./src/plugins/mapset.ts", "./src/plugins/patches.ts", "src/types/globals.d.ts"]
 }
